+++ src/client/Microsoft.Identity.Client/MsalServiceExceptionFactory.cs
@@ -79,7 +79,7 @@ namespace Microsoft.Identity.Client
             ex.Claims = oAuth2Response?.Claims;
             ex.CorrelationId = oAuth2Response?.CorrelationId;
             ex.SubErrorForLogging = oAuth2Response?.SubError;
-            ex.ErrorCodes = oAuth2Response?.ErrorCodes;
+            ex.ErrorCodesForLogging = oAuth2Response?.ErrorCodes is { } errorCodes ? Array.AsReadOnly(errorCodes) : null;
 
             return ex;
         }